upload file corrections: fetcher in utils, delete extra component, refactor uploadDialog
