+++ src/components/tableToolBar/uploadDialog.jsx
@@ -1,4 +1,4 @@
-import { useEffect, useState, useRef, useCallback } from 'react';
+import { useState, useRef } from 'react';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
@@ -6,29 +6,31 @@ import DialogTitle from '@material-ui/core/DialogTitle';
 import Button from '@material-ui/core/Button';
 import Input from '@material-ui/core/Input';
 import LinearProgress from '@material-ui/core/LinearProgress';
-import useUploadFile from '../../hooks/useUploadFile.jsx';
 
-function UploadingFile(props) {
-  let dummyToken = localStorage.getItem('token');
-  const { data, error } = useUploadFile({
-    model_name: props.model_name,
-    token: dummyToken,
-    file: props.file,
-  });
-
-  useEffect(() => {
-    if (data) {
-      props.handleFileUploaded();
-    }
-  }, [data]); //[data, error]);
-  return <LinearProgress />;
-}
+import useSWR from 'swr';
+import { fetcherUpload } from '../../utils/fetcher';
 
 export default function UploadDialog(props) {
+  let dummyToken = localStorage.getItem('token');
+  let model_name = '';
+  let query = `mutation {bulkAdd${model_name} }`;
+  const file = useRef(null);
   const [open, setOpen] = useState(true);
   const [fileChosen, setFileChosen] = useState(false);
   const [loading, setLoading] = useState(false);
-  const file = useRef(null);
+
+  const onUploadSucces = (data_d) => {
+    setLoading(false);
+    setOpen(false);
+    file.current = null;
+    props.handleDone();
+  };
+
+  const { error } = useSWR(
+    loading && query ? [query, dummyToken, file] : null,
+    fetcherUpload,
+    { onSuccess: onUploadSucces }
+  );
 
   const handleOnChange = (event) => {
     if (event.target.files.length > 0) {
@@ -43,17 +45,9 @@ export default function UploadDialog(props) {
     setLoading(true);
   };
 
-  const handleFileUploaded = () => {
-    setLoading(false);
-    file.current = null;
-    setOpen(false);
-    props.handleDone();
-  };
-
   const handleOnClose = () => {
     setOpen(false);
   };
-  const dummy_model_name = 'User';
 
   return (
     <>
@@ -79,13 +73,7 @@ export default function UploadDialog(props) {
             Upload
           </Button>
         </DialogActions>
-        {loading && (
-          <UploadingFile
-            file={file.current}
-            model_name={dummy_model_name}
-            handleFileUploaded={handleFileUploaded}
-          />
-        )}
+        {loading && <LinearProgress />}
       </Dialog>
     </>
   );
